refactor: cart 타입 중복 제거 및 일관성 유지
