+++ src/app/cart/page.tsx
@@ -4,21 +4,17 @@ import React from 'react';
 import { NoFooterLayout } from '@/components/layout/layouts';
 import { CartItem as CartItemComponent, CartSelectAll, CartSummary } from '@/components/cart';
 import { useCart } from '@/hooks/useCart';
-import { mockCartData, calculateCartSummary, type CartItem } from '@/data/cart';
+import { mockCartData, calculateCartSummary } from '@/data/cart';
+import type { CartItem } from '@/types/cart';
 
 // 빈 장바구니 컴포넌트
 function EmptyCart() {
   return (
-    <NoFooterLayout>
-      <div className="container mx-auto max-w-4xl px-6 py-8 md:px-8 md:py-12">
-        <h1 className="mb-8 text-center text-2xl font-semibold text-text-100 md:text-3xl">
-          장바구니
-        </h1>
-        <div className="flex flex-col items-center justify-center py-16">
-          <p className="text-lg text-text-200 md:text-xl">장바구니가 비어있습니다.</p>
-        </div>
-      </div>
-    </NoFooterLayout>
+    <div className="flex flex-col items-center justify-center py-16">
+      <div className="mb-4 text-6xl">🛒</div>
+      <h2 className="mb-2 text-xl font-semibold text-text-100">장바구니가 비어있습니다</h2>
+      <p className="text-text-200">상품을 추가해보세요!</p>
+    </div>
   );
 }
