+++ src/types/cart.ts
@@ -1,15 +1,14 @@
-export interface CartProduct {
-  id: string;
-  name: string;
-  mainImage: string;
-  price: number;
-  originalPrice: number;
-  discountRate: number;
-}
+import type { Product } from './product';
+
+// CartProduct를 Product의 서브셋으로 정의
+export type CartProduct = Pick<
+  Product,
+  'id' | 'name' | 'mainImage' | 'price' | 'originalPrice' | 'discountRate'
+>;
 
 export interface CartItem {
   id: string;
-  product: CartProduct;
+  product: Product;
   selectedOption: {
     value: string;
     label: string;
